ExceptionHandler with error customization

+++ src/main/java/com/DevHugo/ProductsApi/services/ProductService.java
@@ -11,25 +11,26 @@ import com.DevHugo.ProductsApi.entity.Product;
 import com.DevHugo.ProductsApi.repositories.ProductsRepository;
 import com.DevHugo.ProductsApi.services.exceptions.ResourceNotFoundException;
 
-
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
-	
+
 	@Autowired
 	private ProductsRepository repo;
-	
+
 	@Transactional(readOnly = true)
-	public Page<ProductDTO> findAll(Pageable pageable){
+	public Page<ProductDTO> findAll(Pageable pageable) {
 		Page<Product> result = repo.findAll(pageable);
-		return result.map(product ->(new ProductDTO(product)));
+		return result.map(product -> (new ProductDTO(product)));
 	}
+
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
 		Product prod = repo.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(prod);
 	}
-	
+
 	@Transactional
 	public ProductDTO insert(ProductDTO dto) {
 		Product entity = new Product();
@@ -39,20 +40,27 @@ public class ProductService {
 		entity = repo.save(entity);
 		return new ProductDTO(entity);
 	}
-	
+
 	@Transactional
 	public ProductDTO update(ProductDTO dto, Long id) {
-		
-		Product entity = repo.getReferenceById(id);
-		entity.setName(dto.getName());
-		entity.setPrice(dto.getPrice());
-		entity.setQuantity(dto.getQuantity());
-		entity = repo.save(entity);
-		return new ProductDTO(entity);
+		try {
+			Product entity = repo.getReferenceById(id);
+			entity.setName(dto.getName());
+			entity.setPrice(dto.getPrice());
+			entity.setQuantity(dto.getQuantity());
+			entity = repo.save(entity);
+			return new ProductDTO(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
-	
+
 	@Transactional
 	public void delete(Long id) {
+	
+		if(!repo.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 		repo.deleteById(id);
 	}
 }
